+++ drivers/timer.c
@@ -7,11 +7,11 @@ static acpi_gas_t timer_block;
 static int extended_timer = 0;
 static int supported = 0;
 
-static uint32_t read_timer_value(){    
+static uint32_t read_timer_value(){
     if(timer_block.address_space == ACPI_GAS_IO){
         return laihost_ind(timer_block.base);
     } else if(timer_block.address_space == ACPI_GAS_MMIO){
-        uint32_t *reg = (uint32_t *)(timer_block.base);
+        uint32_t *reg = (uint32_t *)((uint32_t) timer_block.base);
         return *reg;
     } else {
         lai_panic("Unknown ACPI Timer address space");
@@ -25,7 +25,7 @@ lai_api_error_t lai_start_pm_timer(){
 
     if(fadt->pm_timer_length != 4)
         return LAI_ERROR_UNSUPPORTED;
-    
+
     supported = 1;
 
     if(lai_current_instance()->acpi_revision >= 2 && fadt->x_pm_timer_block.base){
@@ -48,7 +48,7 @@ lai_api_error_t lai_start_pm_timer(){
 lai_api_error_t lai_stop_pm_timer(){
     if(!supported)
         return LAI_ERROR_UNSUPPORTED;
-    
+
     lai_set_sci_event(lai_get_sci_event() & ~ACPI_TIMER);
     return LAI_ERROR_NONE;
 }
@@ -56,7 +56,7 @@ lai_api_error_t lai_stop_pm_timer(){
 lai_api_error_t lai_busy_wait_pm_timer(uint64_t ms){
     if(!supported)
         return LAI_ERROR_UNSUPPORTED;
-    
+
     // number of ticks per millisecond 3579.545, rounded up to 3580
     uint32_t goal = read_timer_value() + (ms * 3580);
